requirements.txt
Flask==2.2.3
Pillow==9.4.0
requests==2.28.2
reportlab==3.6.12
Werkzeug==2.2.3
